refactor: 필요없는 설정 정리
