--- src/main/java/in/koreatech/koin/global/config/RestTemplateConfig.java
@@ -1,26 +1,0 @@
-package in.koreatech.koin.global.config;
-
-import static java.nio.charset.StandardCharsets.UTF_8;
-
-import java.time.Duration;
-
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.client.BufferingClientHttpRequestFactory;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
-import org.springframework.http.converter.StringHttpMessageConverter;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestTemplateConfig {
-
-    @Bean
-    public RestTemplate restTemplate(RestTemplateBuilder restTemplateBuilder) {
-        return restTemplateBuilder.
-            requestFactory(() -> new BufferingClientHttpRequestFactory(new SimpleClientHttpRequestFactory())).
-            setConnectTimeout(Duration.ofMillis(5000))
-            .setReadTimeout(Duration.ofMillis(5000))
-            .additionalMessageConverters(new StringHttpMessageConverter(UTF_8)).build();
-    }
-}
